Extract common logic into an abstract class
